fix(tui): preserve native scrollback and scheme-aware contrast
Track history progress by source grapheme positions across rewraps and partial-line updates. Use LF insertion with aligned Unicode widths, defer clipped rows, and restore the viewport in a bounded draw cycle.

Separate canvas and dark-surface colors, add regression coverage, and validate complete terminal output, styles, composer and cursor in CI.

+++ crates/tui/src/app/view/ui/layout.rs
@@ -73,7 +73,7 @@ pub(crate) fn desired_height(app: &mut AppState, width: u16, height: u16) -> u16
 
     let max_log_height = remaining_height.saturating_sub(reserved_height);
     let wrapped_total = cached_wrap_log_lines(app, width as usize).len();
-    let mut desired_log_height = (wrapped_total as u16).min(max_log_height);
+    let mut desired_log_height = wrapped_total.min(usize::from(max_log_height)) as u16;
     if desired_log_height == 0 && max_log_height > 0 && wrapped_total > 0 {
         desired_log_height = 1;
     }
